refactor(plugins): separate DatabaseType from the registry it asks
DatabaseType is a string a connection carries and a value the display layer
compares against a constant, but its 34 registry-backed members were declared
beside it, so naming a database type anywhere cost the plugin registry, the
snapshot table and the theme engine at compile time.

The value, its constants and its Codable conformance move to DatabaseType.swift
with no dependencies; the lookups move to DatabaseType+Registry.swift next to
the registry they ask. Nothing else changes: same members, same call sites.

CellDisplayFormatter's compile closure, 46 files after the display layer was
cut from the preference layer, is now 10.
